requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
requests==2.32.3
httpx==0.27.2
qdrant-client==1.9.1
openai==1.40.3
pymupdf==1.24.9
pdfplumber==0.11.4
python-docx==1.1.2
pytesseract==0.3.13
Pillow==10.4.0
langdetect==1.0.9
jinja2==3.1.4
rapidfuzz==3.9.6
tqdm==4.66.5
